+++ src/scripts/app/cms/cms.html
@@ -2,7 +2,7 @@
   <h1>Create Grammar Materials</h1>
   <ul>
     <li><a ui-sref="cms-grammar-activities()">Create Grammar Activities (Sentence Writing)</a></li>
-    <li><a ui-sref="cms-proofreading-activities()">Create Proofreading Activities (Passages)</a></li>
+    <li><a ui-sref="cms-proofreader-activities()">Create Proofreader Activities (Passages)</a></li>
     <li><a ui-sref="cms-concepts()">Create Grammar Concepts (Rules)</a></li>
   </ul>
 </div>
