+++ src/scripts/app/cms/cms.module.js
@@ -5,7 +5,7 @@ angular
   .module('quill-grammar.cms', [
     require('./concepts/module.js').name,
     require('./grammarActivities/module.js').name,
-    require('./proofreadingActivities/module.js').name,
+    require('./proofreaderActivities/module.js').name,
   ])
   .config(require('./cms.config.js'))
   .controller('cms', require('./cms.controller.js'));
