+++ src/scripts/app/cms/proofreaderActivities/controller.js
@@ -3,7 +3,7 @@
 module.exports =
 
 /*@ngInject*/
-function ProofreadingActivitiesCmsCtrl (
+function ProofreaderActivitiesCmsCtrl (
 
 ) {
 };
