+++ src/scripts/app/cms/proofreaderActivities/module.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports =
+
+angular.module('quill-grammar.cms.proofreaderActivities', [
+
+])
+.config(require('./config.js'))
+.controller('ProofreaderActivitiesCmsCtrl', require('./controller.js'))
+;
