+++ src/scripts/app/cms/proofreaderActivities/proofreader.activities.cms.html
@@ -1,0 +1,1 @@
+<h3>Proofreader Activities</h3>
