--- src/scripts/app/cms/proofreadingActivities/module.js
@@ -1,10 +1,0 @@
-'use strict';
-
-module.exports =
-
-angular.module('quill-grammar.cms.proofreadingActivities', [
-
-])
-.config(require('./config.js'))
-.controller('ProofreadingActivitiesCmsCtrl', require('./controller.js'))
-;
