--- src/scripts/app/cms/proofreadingActivities/proofreading.activities.cms.html
@@ -1,1 +1,0 @@
-<h3>Proofreading Activities</h3>
